# api/proto/web/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: "3"
paths:
    /api/v1/auth/email-login:
        post:
            tags:
                - Auth
            description: 邮箱验证码登录接口
            operationId: Auth_EmailLogin
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AuthEmailLoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AuthEmailLoginResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/auth/forget:
        post:
            tags:
                - Auth
            description: 找回密码接口
            operationId: Auth_Forget
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AuthForgetRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AuthForgetResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/auth/login:
        post:
            tags:
                - Auth
            description: 登录接口
            operationId: Auth_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AuthLoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AuthLoginResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
    /api/v1/auth/oauth:
        post:
            tags:
                - Auth
            description: 获取 oauth2.0 跳转地址
            operationId: Auth_Oauth
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AuthOauthRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AuthOauthResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/auth/oauth/bind:
        post:
            tags:
                - Auth
            description: 绑定第三方登录接口
            operationId: Auth_OauthBind
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AuthOAuthBindRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AuthOAuthBindResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/auth/oauth/login:
        post:
            tags:
                - Auth
            description: 第三方登录接口
            operationId: Auth_OauthLogin
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AuthOauthLoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AuthOauthLoginResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/auth/refresh-token:
        post:
            tags:
                - Auth
            description: Token 刷新接口
            operationId: Auth_RefreshToken
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AuthRefreshTokenRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AuthRefreshTokenResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/auth/register:
        post:
            tags:
                - Auth
            description: 注册接口
            operationId: Auth_Register
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AuthRegisterRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AuthRegisterResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/common/send-email:
        post:
            tags:
                - Common
            description: 发送邮件验证码接口
            operationId: Common_SendEmail
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CommonSendEmailRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CommonSendEmailResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/common/send-sms:
        post:
            tags:
                - Common
            description: |-
                发送短信验证码接口
                    @Summary 发送短信验证码111222
            operationId: Common_SendSms
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CommonSendSmsRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CommonSendSmsResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/common/send-test:
        post:
            tags:
                - Common
            description: 俺们就开始的那
            operationId: Common_Test
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CommonSendTestRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CommonSendTestResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
components:
    schemas:
        AuthEmailLoginRequest:
            required:
                - email
                - emailCode
                - platform
            type: object
            properties:
                email:
                    type: string
                    description: 邮箱
                emailCode:
                    type: string
                    description: 邮箱验证码
                platform:
                    type: string
                    description: 登录平台
            description: 邮箱登录请求参数
        AuthEmailLoginResponse:
            required:
                - type
                - accessToken
                - expiresIn
            type: object
            properties:
                type:
                    type: string
                    description: Token 类型
                accessToken:
                    type: string
                    description: token
                expiresIn:
                    type: integer
                    description: 过期时间
                    format: int32
            description: 邮箱登录响应参数
        AuthForgetRequest:
            required:
                - email
                - password
                - emailCode
            type: object
            properties:
                email:
                    type: string
                    description: 邮箱
                password:
                    type: string
                    description: 登录密码
                emailCode:
                    type: string
                    description: 邮箱验证码
            description: 找回密码接口请求参数
        AuthForgetResponse:
            type: object
            properties: {}
            description: 找回密码接口响应参数
        AuthLoginRequest:
            required:
                - mobile
                - password
                - platform
            type: object
            properties:
                mobile:
                    type: string
                    description: 登录账号（字段名仍为 mobile）：仅与 users.username 一致
                password:
                    type: string
                    description: 登录密码
                platform:
                    type: string
                    description: 登录平台
            description: 登录接口请求参数
        AuthLoginResponse:
            required:
                - type
                - accessToken
                - expiresIn
                - isTrans
            type: object
            properties:
                type:
                    type: string
                    description: Token 类型
                accessToken:
                    type: string
                    description: token
                expiresIn:
                    type: integer
                    description: 过期时间
                    format: int32
                isTrans:
                    type: integer
                    description: 是否已设置交易密码（0/1）
                    format: int32
            description: 登录接口响应参数
...
        AuthOAuthBindRequest:
            required:
                - bindToken
                - mobile
                - smsCode
            type: object
            properties:
                bindToken:
                    type: string
                mobile:
                    type: string
                smsCode:
                    type: string
        AuthOAuthBindResponse:
            required:
                - authorize
            type: object
            properties:
                authorize:
                    $ref: '#/components/schemas/Authorize'
        AuthOauthLoginRequest:
            required:
                - oauthType
                - code
                - state
            type: object
            properties:
                oauthType:
                    type: string
                code:
                    type: string
                state:
                    type: string
        AuthOauthLoginResponse:
            required:
                - isAuthorize
                - authorize
                - bindToken
            type: object
            properties:
                isAuthorize:
                    type: string
                authorize:
                    $ref: '#/components/schemas/Authorize'
                bindToken:
                    type: string
        AuthOauthRequest:
            required:
                - oauthType
            type: object
            properties:
                oauthType:
                    type: string
        AuthOauthResponse:
            required:
                - uri
            type: object
            properties:
                uri:
                    type: string
        AuthRefreshTokenRequest:
            type: object
            properties: {}
            description: Token 刷新接口请求参数
        AuthRefreshTokenResponse:
            required:
                - type
                - accessToken
                - expiresIn
            type: object
            properties:
                type:
                    type: string
                    description: Token 类型
                accessToken:
                    type: string
                    description: token
                expiresIn:
                    type: integer
                    description: 过期时间
                    format: int32
            description: Token 刷新接口响应参数
        AuthRegisterRequest:
            type: object
            properties:
                nickname:
                    type: string
                    description: 昵称
                mobile:
                    type: string
                    description: 手机号（可选）
                email:
                    type: string
                    description: 邮箱（可选）
                password:
                    type: string
                    description: 登录密码
                platform:
                    type: string
                    description: 登录平台
                smsCode:
                    type: string
                    description: 短信验证码（手机注册时需要）
                inviteCode:
                    type: string
                    description: 邀请码（可选）
                emailCode:
                    type: string
                    description: 邮箱验证码（邮箱注册时需要）
            description: 注册接口请求参数
        AuthRegisterResponse:
            required:
                - type
                - accessToken
                - expiresIn
            type: object
            properties:
                type:
                    type: string
                    description: Token 类型
                accessToken:
                    type: string
                    description: token
                expiresIn:
                    type: integer
                    description: 过期时间
                    format: int32
            description: 注册接口响应参数
...
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
...